move more things to crosswords
