+++ projects/backend/fronts.ts
@@ -1,12 +1,5 @@
 import { s3fetch, s3Latest } from './s3'
-import {
-    Front,
-    Collection,
-    CAPIArticle,
-    Crossword,
-    Appearance,
-    cardLayouts,
-} from './common'
+import { Front, Collection, CAPIArticle, Appearance } from './common'
 import { LastModifiedUpdater } from './lastModified'
 import {
     attempt,
@@ -25,11 +18,7 @@ import {
     PublishedFurtniture,
     PublishedFront,
 } from './fronts/issue'
-import {
-    getCrosswordType,
-    getCrosswordName,
-    getCrosswordKicker,
-} from './utils/crossword'
+import { getCrosswordArticleOverrides } from './utils/crossword'
 
 export const parseCollection = async (
     collectionResponse: PublishedCollection,
@@ -78,18 +67,12 @@ export const parseCollection = async (
 
             switch (article.type) {
                 case 'crossword':
-                    const type = getCrosswordType(article.path)
                     return [
                         article.path,
                         {
                             ...article,
+                            ...getCrosswordArticleOverrides(article),
                             key: article.path,
-                            headline: getCrosswordName(type),
-                            kicker: getCrosswordKicker(article.crossword),
-                            image: getImageFromURL(
-                                'https://media.guim.co.uk/3671bfc12549d3ebac611f96eb0dc234a620e008/0_41_5232_3139/5232.jpg',
-                            ),
-                            crossword: (article.crossword as unknown) as Crossword,
                         },
                     ]
